requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.0.5
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
Bottleneck==1.4.0
certifi==2024.12.14
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
executing==0.8.3
filelock==3.13.1
FinMind==1.7.8
Flask==3.1.0
fonttools==4.54.1
frozendict==2.4.4
frozenlist==1.5.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.7.2
kiwisolver==1.4.4
loguru==0.7.3
lxml==5.2.2
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.3
numexpr==2.8.7
numpy==1.26.4
packaging==24.1
pandas==2.2.3
parso==0.8.3
peewee==3.17.6
pexpect==4.8.0
pillow==10.4.0
pip==24.2
platformdirs==3.10.0
pluggy==1.5.0
prettytable==3.12.0
prompt-toolkit==3.0.43
propcache==0.2.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.10.4
pydantic_core==2.27.2
pyecharts==2.0.7
Pygments==2.15.1
pyparsing==3.1.4
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
requests==2.32.3
scikit-learn==1.5.1
scipy==1.13.1
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
soupsieve==2.6
stack-data==0.2.0
sympy==1.13.3
ta==0.5.25
TA-Lib==0.4.32
threadpoolctl==3.5.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.44.0
xgboost==2.1.1
yarl==1.18.3
yfinance==0.2.43
zipp==3.21.0
